Patched to compile C67
Patched tcc0.9.26 to generate correct C67 coff files on Darwin X86_64.
Have not tried other platforms yet. Patches was not only for
architechure but bugs as well.

+++ KMotionX/tcc-0.9.26/tcc.h
@@ -294,8 +294,8 @@ typedef union CValue {
     int i;
     unsigned int ui;
     unsigned int ul; /* address (should be unsigned long on 64 bit cpu) */
-    long long ll;
-    unsigned long long ull;
+    int64_t ll;
+    uint64_t ull;
     struct CString *cstr;
     void *ptr;
     int tab[LDOUBLE_SIZE/4];
